A stored app reaches the callback, the consent URL, and the strand guard
Three ways an app configured in Settings rather than the environment failed to
be the app that actually got used:

The callback resolved its app under the raw request context. That route is
session-less by design, so it carries no workspace, and a per-workspace stored
app is invisible from it — the code was spent against the deployment's client
or against nothing. It resolves under the signed grantor context now.

The signer and the callback bases were set only when this seam built the
registry. Google's wiring builds it first whenever both are configured, so
Microsoft consent signed no state and addressed an empty base. They are what
consent needs rather than what the registry needs, and are set independently.

The strand guard counted only 'connected' connections. An errored connection
still holds the refresh token the old client issued and the sweep still tries
it, so changing the id stranded exactly the installation most likely to be
changing it — one whose mailboxes had started failing.
