+++ backend/internal/compose/integration/connectorapp_integration_test.go
@@ -188,93 +188,6 @@ func TestRotatingTheGoogleAppRetiresTheOldSecret(t *testing.T) {
 	assertVaultRefGone(ctx, t, e, firstRef, "the superseded client secret")
 }
 
-// A CLIENT ID CANNOT BE REPLACED WHILE MAILBOXES ARE CONNECTED UNDER IT.
-//
-// A refresh token belongs to the client that issued it. Swapping the id makes
-// every stored token unrefreshable, and the vendor answers `invalid_client` — so
-// the mailboxes stop syncing one by one, at whatever hour their next refresh
-// falls, and from the inside it looks like every mailbox revoking access at
-// once.
-//
-// The SECRET stays rotatable against the same id, which is the whole point of
-// being able to rotate one.
-func TestReplacingTheClientIDIsRefusedWhileConnectionsHoldTokensFromTheOldOne(t *testing.T) {
-	e := SetupSearch(t)
-	ctx := e.captureAdmin()
-	store := googleAppStore(t, e)
-
-	if err := store.Set(ctx, capture.AppProviderGoogle, testClientID, "GOCSPX-first", ""); err != nil {
-		t.Fatalf("storing: %v", err)
-	}
-
-	// A ROTATION FIRST, before any connection exists, so the refusal below is
-	// about the id and not about the write path being closed.
-	if err := store.Set(ctx, capture.AppProviderGoogle, testClientID, "GOCSPX-second", ""); err != nil {
-		t.Fatalf("rotating the secret with no connections: %v", err)
-	}
-
-	seedGmailConnection(ctx, t, e)
-
-	// The secret still rotates with a mailbox connected — same client, same
-	// tokens.
-	if err := store.Set(ctx, capture.AppProviderGoogle, testClientID, "GOCSPX-third", ""); err != nil {
-		t.Fatalf("rotating the secret with a connection: %v", err)
-	}
-
-	const otherClientID = "9999999999-zyxwvutsrqponmlk.apps.googleusercontent.com"
-	err := store.Set(ctx, capture.AppProviderGoogle, otherClientID, "GOCSPX-fourth", "")
-	var invalid settings.InvalidValue
-	if !errors.As(err, &invalid) {
-		t.Fatalf("replacing the client id answered %v, want a refusal — every connected mailbox holds a "+
-			"refresh token the new client cannot use", err)
-	}
-	if !strings.Contains(invalid.Reason, "refresh token") {
-		t.Errorf("the refusal reads %q, which does not tell the operator why", invalid.Reason)
-	}
-
-	// And the stored app is untouched: a refused write that half-committed
-	// would be worse than the change it refused.
-	app, _, err := store.Credentials(ctx, capture.AppProviderGoogle)
-	if err != nil {
-		t.Fatalf("resolving after the refusal: %v", err)
-	}
-	if app.ClientID != testClientID {
-		t.Errorf("client id = %q after a refused change, want the one that was there", app.ClientID)
-	}
-	if app.ClientSecretRef != "GOCSPX-third" {
-		t.Errorf("secret = %q after a refused change, want the one that was there", app.ClientSecretRef)
-	}
-
-	// A DISCONNECTED mailbox strands nobody, so the change goes through once
-	// the operator has done what the refusal asked.
-	if err := database.WithWorkspaceTx(e.Admin(), e.Pool, func(tx pgx.Tx) error {
-		_, err := tx.Exec(context.Background(),
-			`UPDATE capture_connection SET status = 'disconnected' WHERE provider = 'gmail'`)
-		return err
-	}); err != nil {
-		t.Fatalf("disconnecting: %v", err)
-	}
-	if err := store.Set(ctx, capture.AppProviderGoogle, otherClientID, "GOCSPX-fifth", ""); err != nil {
-		t.Fatalf("replacing the client id with nothing connected: %v — the refusal has no way out", err)
-	}
-}
-
-// seedGmailConnection writes one connected gmail mailbox. Written directly
-// because the subject is the ROW the app change would strand, not how a consent
-// callback puts it there.
-func seedGmailConnection(ctx context.Context, t *testing.T, e *SearchEnv) {
-	t.Helper()
-	if err := database.WithWorkspaceTx(e.Admin(), e.Pool, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			INSERT INTO capture_connection (provider, user_id, scopes, status, auth)
-			VALUES ('gmail', $1, '{}', 'connected', $2)`,
-			e.Rep1, []byte(`{"refresh_token":"r","granted":[]}`))
-		return err
-	}); err != nil {
-		t.Fatalf("seeding the gmail connection: %v", err)
-	}
-}
-
 // Removing clears the app and destroys the secret, and removing an absent one
 // succeeds: the caller asked for a state and that state already holds.
 func TestRemovingTheGoogleAppIsIdempotentAndDestroysTheSecret(t *testing.T) {
